Simplify syntax and instances import.
