+++ examples/features/traverse/TraverseWs.sc
@@ -1,33 +1,33 @@
-import advxml.instances.traverse._
-import advxml.syntax.traverse.try_._
-import cats.instances.option._
 
-import scala.xml.NodeSeq
+object Test {
 
-//TODO: TO Fix location
-import advxml.instances.transform.predicates._
+  import cats.instances.option._
+  import scala.xml.NodeSeq
+  import advxml.implicits._
 
-val document =
-  <Orders>
-    <Order Id="1">
-      <OrderLines>
-        <OrderLine Id="1" Price="100€"/>
-        <OrderLine Id="2" Price="50€"/>
-      </OrderLines>
-    </Order>
-    <Order Id="2">
-      <OrderLines>
-        <OrderLine Id="1" Price="10€"/>
-        <OrderLine Id="2" Price="20€"/>
-      </OrderLines>
-    </Order>
-  </Orders>
+  val document =
+    <Orders>
+      <Order Id="1">
+        <OrderLines>
+          <OrderLine Id="1" Price="100€"/>
+          <OrderLine Id="2" Price="50€"/>
+        </OrderLines>
+      </Order>
+      <Order Id="2">
+        <OrderLines>
+          <OrderLine Id="1" Price="10€"/>
+          <OrderLine Id="2" Price="20€"/>
+        </OrderLines>
+      </Order>
+    </Orders>
 
-val order2Opt: Option[NodeSeq] = (document \? "Order" \? "OrderLines" \? "OrderLine")
-    .map(_.filter(attrs(("Id", _ == "2"))))
+  val order2Opt: Option[NodeSeq] = (document \? "Order" \? "OrderLines" \? "OrderLine")
+    .map(_.filter(attrs(k"Id" === "2")))
 
-val order2OptUsingDynamic: Option[NodeSeq] = document.\?*.Order.\?*.OrderLines.\?*.OrderLine.get
-    .map(_.filter(attrs(("Id", _ == "2")))).>>.atIndex(0)
+  val order2OptUsingDynamic: Option[NodeSeq] = document.\?*.Order.OrderLines.get
+    .filterChild(attrs(k"Id" === "2"))
+}
 
+Test.order2Opt
 
-val order2OptUsingDynamic = document.\?*.Order.OrderLines.OrderLine.get.childT.filter(attrs(("Id", _ == "2")))
+Test.order2OptUsingDynamic
