+++ src/test/scala/advxml/instances/XmlPredicateTest.scala
@@ -6,7 +6,6 @@ import scala.xml.{Elem, NodeSeq}
 
 class XmlPredicateTest extends AnyFunSuite {
 
-  import advxml.instances.convert._
   import advxml.instances.transform.predicates._
   import advxml.syntax._
   import advxml.testUtils.ScalacticXmlEquality._
